Bug humidity is int (#3)
* Create a config for the github issue template

* Id in the devicesummary is mandatory

* Renamed exception, made devicesummary id mandatory

* Fixed ruff issues

* Fix datatype for humidity

+++ src/pyfreshr/models.py
@@ -56,7 +56,7 @@ class DeviceReadings:
     t4: float | None = None
     flow: float | None = None
     co2: int | None = None
-    hum: float | None = None
+    hum: int | None = None
     dp: float | None = None
     temp: float | None = None
     extras: dict[str, Any] = field(default_factory=dict)
